added dropdowns for conditions & moved api
