+++ modules/ndu/src/backend/conditions.ts
@@ -7,7 +7,15 @@ export const dialogConditions: sdk.Condition[] = [
     label: 'User is using a specific channel',
     description: `The user speaks on channel {channelName}`,
     params: {
-      channelName: { label: 'Name of the channel', type: 'string' }
+      channelName: {
+        label: 'Select a channel from the list',
+        type: 'list',
+        list: {
+          endpoint: 'BOT_API_PATH/mod/ndu/channels',
+          valueField: 'value',
+          labelField: 'label'
+        }
+      }
     },
     evaluate: (event, params) => {
       return event.channel === params.channelName ? 1 : 0
@@ -43,7 +51,18 @@ export const dialogConditions: sdk.Condition[] = [
   {
     id: 'user_topic_source',
     label: 'User is coming from a specific topic',
-    params: { topicName: { label: 'Name of the topic', type: 'string' } },
+    description: `The user's last topic was {topicName}`,
+    params: {
+      topicName: {
+        label: 'Name of the topic',
+        type: 'list',
+        list: {
+          endpoint: 'BOT_API_PATH/topics',
+          valueField: 'name',
+          labelField: 'name'
+        }
+      }
+    },
     evaluate: (event, params) => {
       const topics = event.state.session.lastTopics
       return topics && topics[topics.length - 1] === params.topicName ? 1 : 0
